requirements.txt
tensorflow==2.15.0
keras==2.15.0
numpy<2.0
protobuf<4
h5py==3.10.0
opt-einsum<4

streamlit==1.32.2
pillow>=10.0.0
requests
opencv-python-headless
pandas
scikit-learn